Remove comments to make code clearer
Comments still have their place, though.

Rather than explaining what the code does, comments should
explain why something is done in a particular way.

The code by itself should be clear enough to express what
it is trying to do, and Ruby is one of the best languages
in that department.